move ingredient PropTypes to constants
